Refactor botpose reset into drivetrain command factory
Move reset logic out of RobotContainer lambdas and into
CommandSwerveDrivetrain.resetPoseFromVisionCommand(). Bindings in
RobotContainer are now single-line calls. Logic and imports
(LimelightHelpers, Constants, Units, SmartDashboard) live with
the drivetrain where they belong.
