+++ src/main/java/frc/robot/RobotContainer.java
@@ -11,8 +11,6 @@ import com.ctre.phoenix6.swerve.SwerveRequest;
 
 import choreo.auto.AutoChooser;
 import choreo.auto.AutoFactory;
-import edu.wpi.first.math.util.Units;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
@@ -28,7 +26,6 @@ import frc.robot.subsystems.intake.IntakeIOHardware;
 import frc.robot.subsystems.shooter.ShooterIOHardware;
 import frc.robot.subsystems.shooter.ShooterSubsystem;
 import frc.robot.subsystems.led.LedSubsystem;
-import frc.robot.subsystems.vision.LimelightHelpers;
 import frc.robot.subsystems.vision.VisionIOLimelight;
 import frc.robot.subsystems.vision.VisionSubsystem;
 
@@ -114,25 +111,7 @@ public class RobotContainer {
         driver.start().onTrue(drivetrain.runOnce(drivetrain::seedFieldCentric));
 
         // Back: Reset odometry to Limelight botpose (use when robot rides up on a ball and wheels lose contact)
-        driver.back().onTrue(Commands.runOnce(() -> {
-            var driveState = drivetrain.getState();
-            double omegaRps = Units.radiansToRotations(driveState.Speeds.omegaRadiansPerSecond);
-            if (Math.abs(omegaRps) >= 2.0) {
-                SmartDashboard.putString("BotposeReset", "SKIPPED (spinning)");
-                return;
-            }
-            double headingDeg = driveState.Pose.getRotation().getDegrees();
-            LimelightHelpers.SetRobotOrientation(
-                    Constants.Vision.LIMELIGHT4_NAME, headingDeg, 0, 0, 0, 0, 0);
-            var llMeasurement = LimelightHelpers.getBotPoseEstimate_wpiBlue_MegaTag2(
-                    Constants.Vision.LIMELIGHT4_NAME);
-            if (llMeasurement != null && llMeasurement.tagCount > 0) {
-                drivetrain.resetPose(llMeasurement.pose);
-                SmartDashboard.putString("BotposeReset", "OK");
-            } else {
-                SmartDashboard.putString("BotposeReset", "SKIPPED (no tags)");
-            }
-        }, drivetrain));
+        driver.back().onTrue(drivetrain.resetPoseFromVisionCommand());
 
         drivetrain.registerTelemetry(logger::telemeterize);
 
@@ -187,25 +166,7 @@ public class RobotContainer {
         operator.leftBumper().whileTrue(intake.retractSlidesStack());
 
         // Back (View ⧉): Reset odometry to botpose — use when robot rides up on a ball
-        operator.back().onTrue(Commands.runOnce(() -> {
-            var driveState = drivetrain.getState();
-            double omegaRps = Units.radiansToRotations(driveState.Speeds.omegaRadiansPerSecond);
-            if (Math.abs(omegaRps) >= 2.0) {
-                SmartDashboard.putString("BotposeReset", "SKIPPED (spinning)");
-                return;
-            }
-            double headingDeg = driveState.Pose.getRotation().getDegrees();
-            LimelightHelpers.SetRobotOrientation(
-                    Constants.Vision.LIMELIGHT4_NAME, headingDeg, 0, 0, 0, 0, 0);
-            var llMeasurement = LimelightHelpers.getBotPoseEstimate_wpiBlue_MegaTag2(
-                    Constants.Vision.LIMELIGHT4_NAME);
-            if (llMeasurement != null && llMeasurement.tagCount > 0) {
-                drivetrain.resetPose(llMeasurement.pose);
-                SmartDashboard.putString("BotposeReset", "OK");
-            } else {
-                SmartDashboard.putString("BotposeReset", "SKIPPED (no tags)");
-            }
-        }, drivetrain));
+        operator.back().onTrue(drivetrain.resetPoseFromVisionCommand());
 
         // FIXME Add a reverse indexer on start button for operator
