+++ src/main/java/frc/robot/subsystems/CommandSwerveDrivetrain.java
@@ -20,15 +20,19 @@ import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
+import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import edu.wpi.first.wpilibj.Notifier;
 import edu.wpi.first.wpilibj.RobotController;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Subsystem;
 import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 
+import frc.robot.Constants;
 import frc.robot.generated.TunerConstants.TunerSwerveDrivetrain;
+import frc.robot.subsystems.vision.LimelightHelpers;
 
 /**
  * Class that extends the Phoenix 6 SwerveDrivetrain class and implements
@@ -297,6 +301,38 @@ public class CommandSwerveDrivetrain extends TunerSwerveDrivetrain implements Su
                 .withRotationalRate(0));
         });
     }
+    /**
+     * Returns a command that resets the robot pose to the Limelight MegaTag2 botpose estimate.
+     * <p>
+     * Use this when the robot rides up on a ball and the wheels lose ground contact,
+     * corrupting wheel odometry. The reset is skipped if the robot is spinning too fast
+     * (>= 2.0 rot/s) — a stale heading fed to MegaTag2 would produce a bad pose estimate.
+     * Result is published to SmartDashboard under "BotposeReset" for drive-team feedback.
+     *
+     * @return Command (requires drivetrain)
+     */
+    public Command resetPoseFromVisionCommand() {
+        return runOnce(() -> {
+            var driveState = getState();
+            double omegaRps = Units.radiansToRotations(driveState.Speeds.omegaRadiansPerSecond);
+            if (Math.abs(omegaRps) >= 2.0) {
+                SmartDashboard.putString("BotposeReset", "SKIPPED (spinning)");
+                return;
+            }
+            double headingDeg = driveState.Pose.getRotation().getDegrees();
+            LimelightHelpers.SetRobotOrientation(
+                    Constants.Vision.LIMELIGHT4_NAME, headingDeg, 0, 0, 0, 0, 0);
+            var llMeasurement = LimelightHelpers.getBotPoseEstimate_wpiBlue_MegaTag2(
+                    Constants.Vision.LIMELIGHT4_NAME);
+            if (llMeasurement != null && llMeasurement.tagCount > 0) {
+                resetPose(llMeasurement.pose);
+                SmartDashboard.putString("BotposeReset", "OK");
+            } else {
+                SmartDashboard.putString("BotposeReset", "SKIPPED (no tags)");
+            }
+        });
+    }
+
     @Override
     public void periodic() {
         /*
